requirements.txt
firebase-admin==6.5.0
python-dotenv==1.0.1
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.4
# After installation run: python -m spacy download de_core_news_sm
